+++ Client/src/component/header/header.css
@@ -26,10 +26,11 @@ header{
     display: inline-block;
     padding: 0px 6px;
     margin: 0 5px;
+    cursor: pointer;
 }
 
-  .right_nav{
-    text-align: right;
-    font-size: 16px;
-    font-weight: 700;
-  }
+.right_nav{
+  text-align: right;
+  font-size: 16px;
+  font-weight: 700;
+}
